geo: Initial geometry library (#45)
* core: Add asin and acos utils

* core: Add f32 clamp routine

* core: Add degress / radian conversions

* core: Add linear interpolate utilities

* geo: Add geometric vector type

* cmake: Enable additional conversion warnings

* geo: Add doc comment to vector

* core: Add tan, atan and atan2 utils

* core: Minor cleanup of static consts

* geo: Add initial quaternion implementation

* geo: Add initial matrix impl

* cmake: Ignore warning 4100 on msvc

Gives false positives

* geo: Add more matrix test cases

* geo: Add util to convert quat to matrix

* geo: Add util to convert matrix to quaternion

* geo: Add a util for look rotations

* geo: Fix msvc compat

* geo: Add axis aligned box type

+++ libs/cli/src/parse.c
@@ -46,7 +46,7 @@ static void cli_parse_consume_arg(CliParseCtx* ctx) {
   ++ctx->argHead;
 }
 
-static u32 cli_parse_args_remaining(CliParseCtx* ctx) { return ctx->argTail - ctx->argHead; }
+static u32 cli_parse_args_remaining(CliParseCtx* ctx) { return (u32)(ctx->argTail - ctx->argHead); }
 
 static bool cli_parse_already_provided(CliParseCtx* ctx, CliId id) {
   return dynarray_at_t(&ctx->options, id, CliInvocationOption)->provided;
